+++ inst/include/epiworld/epiworld.hpp
@@ -127,7 +127,7 @@ namespace epiworld {
     #include "tools/vaccine.hpp"
     #include "globalevents/quarantinetrigger-meat.hpp"
     #include "globalevents/bubbles-meat.hpp"
-    
+
     #include "models/models.hpp"
 
 }